test refactored
